+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## ?.?.? - Unreleased
+## 4.2.0 - 2024-06-18
 
 * New commands `okdata -e` and `okdata -v` for printing the current environment
   and the current version of okdata-cli, respectively.